utils - searchBar + added renderChildren to folderInfo util
