+++ src/js/bookmarkManager/searchBar.js
@@ -1,3 +1,13 @@
+import {
+  renderChildren,
+  showFolderInfo,
+  showFolderInfoDisplay
+} from '../utils/folderInfo';
+import { hideBookmarkInfo } from '../utils/bookmarkInfo';
+import { hideFolderInfoEdit } from '../utils/managerForm';
+
+//depends: globalSelectHandler
+
 document.addEventListener('DOMContentLoaded', () => {
   const searchBar = document.getElementById('search-input');
   searchBar.addEventListener('keyup', () => {
@@ -9,8 +19,3 @@ document.addEventListener('DOMContentLoaded', () => {
     globalSelectHandler.unselect();
   });
 });
-
-function clearSearchBar() {
-  document.getElementById('search-input').value = '';
-  M.updateTextFields();
-}
