+++ src/js/utils/folderInfo.js
@@ -48,7 +48,12 @@ export function findLeafNodes(rootNode, onLeafNodeFound) {
 }
 
 import options from '../config/config';
-const { openedArrowIcon, openedFolderIcon } = options;
+const {
+  openedArrowIcon,
+  openedFolderIcon,
+  defaultFileIconColor,
+  trackedFileIconColor
+} = options;
 
 //depends: openedArrowIcon, openedFolderIcon
 export function openFolder(folderId) {
@@ -87,4 +92,65 @@ export function openAllParentFolders(parentId) {
 // converts bookmarkId into child-info id
 export function createChildInfoId(bookmarkId) {
   return `child-info-${bookmarkId}`;
+}
+
+/* depends: 
+ hideFolderInfoChildren, findLeafNodes,
+ defaultFileIconColor, trackedFileIconColor
+*/
+// renders children based on search pattern from 'search-input'
+export function renderChildren(renderAll = false) {
+  const childrenList = document.getElementById('folder-children-info');
+  const folderId =
+    renderAll === true ? '0' : childrenList.getAttribute('folderId');
+
+  let searchPattern;
+  try {
+    let searchInput = document.getElementById('search-input').value;
+    searchPattern = new RegExp(searchInput, 'i');
+  } catch {
+    console.warn('invalid reg exp');
+    searchPattern = new RegExp(); // this matched anything
+  }
+
+  chrome.bookmarks.getSubTree(folderId, (results) => {
+    if (chrome.runtime.lastError) {
+      console.warn(chrome.runtime.lastError.message);
+    } else {
+      chrome.storage.sync.get(['dynBookmarks'], ({ dynBookmarks }) => {
+        const dynBook = dynBookmarks || {};
+        hideFolderInfoChildren();
+        for (let child of results) {
+          findLeafNodes(child, (node) => {
+            const childEl = document.getElementById(`child-info-${node.id}`);
+            if (childEl) {
+              if (searchPattern.test(childEl.textContent)) {
+                childEl.parentElement.classList.remove('hide');
+              }
+              const spans = childEl.querySelectorAll('span');
+              for (let span of spans) {
+                if (dynBook[node.id]) {
+                  span.classList.replace(
+                    defaultFileIconColor,
+                    trackedFileIconColor
+                  );
+                } else {
+                  span.classList.replace(
+                    trackedFileIconColor,
+                    defaultFileIconColor
+                  );
+                }
+              }
+              if (spans[0].textContent !== node.title) {
+                spans[0].textContent = node.title;
+              } else if (childEl.getAttribute('href') !== node.url) {
+                childEl.setAttribute('href', node.url);
+                spans[1].textContent = ` (${node.url})`;
+              }
+            }
+          });
+        }
+      });
+    }
+  });
 }
