+++ src/js/utils/searchBar.js
@@ -1,0 +1,4 @@
+export function clearSearchBar() {
+  document.getElementById('search-input').value = '';
+  M.updateTextFields();
+}
